🚸 error handling for test creation
