+++ src/db/schema.ts
@@ -1,4 +1,5 @@
 import { integer, pgTable, serial, text } from "drizzle-orm/pg-core";
+import { z } from "zod";
 
 export const tests = pgTable("tests", {
   id: serial("id").primaryKey(),
@@ -31,3 +32,59 @@ export type Test = typeof tests.$inferSelect;
 export type NewTest = typeof tests.$inferInsert;
 export type Question = typeof questions.$inferSelect;
 export type NewQuestion = typeof questions.$inferInsert;
+
+// TODO: check if answer in options
+const questionSchema = z.discriminatedUnion("type", [
+  z.object({
+    question: z
+      .string()
+      .min(2, { message: "Topic should be at least 2 characters long." })
+      .max(50, { message: "Topic is too long." }),
+    image: z.string().url().optional(),
+    type: z.literal("multiple"),
+    answer: z.array(
+      z
+        .string()
+        .min(1, {
+          message: "Answer should be at least 1 character long.",
+        })
+        .max(100, { message: "Answer is too long." })
+    ),
+    options: z
+      .array(z.string())
+      .min(2, { message: "There should be at least 2 option." }),
+  }),
+  z.object({
+    question: z
+      .string()
+      .min(2, { message: "Topic should be at least 2 characters long." })
+      .max(50, { message: "Topic is too long." }),
+    image: z.string().url().optional(),
+    type: z.literal("open"),
+  }),
+  z.object({
+    question: z
+      .string()
+      .min(2, { message: "Topic should be at least 2 characters long." })
+      .max(50, { message: "Topic is too long." }),
+    image: z.string().url().optional(),
+    type: z.literal("single"),
+    answer: z.array(
+      z
+        .string()
+        .min(1, {
+          message: "Answer should be at least 1 character long.",
+        })
+        .max(100, { message: "Answer is too long." })
+    ),
+    options: z
+      .array(z.string())
+      .min(2, { message: "There should be at least 2 option." }),
+  }),
+]);
+
+export const questionsSchema = z.object({
+  questions: z
+    .array(questionSchema)
+    .min(1, { message: "The test should have at least one question." }),
+});
